Add some numpy magic to cubehelix
Add general code improvements to the cubehelix math function.

Renamed some variables to be more expressive:
- angle -> phi
- fract -> lambda

This change also disambiguates when fract was used to refer to just
lambda and lambda^gamma.

Made the choice between a scalar saturation and a vector of
saturations a bit cleaner.

Numpy magic:

- Use np.dot to do equation 2. This means the function returns a (n, 3)
  numpy array. The Palettable class seems happy with this though.
  Also apply scaling to 255 right in the rgb equation
- Use np.clip for the clipping

Question of phi equation:

The equation for phi in Green 2011 does not have an extra `+1`
but the Fortran code does, as does the original cubehelix.py
I'm leaving out the +1 to keep to the original equation, but
worth investigating. In practise I see no difference!