Store RGB, class method generates colors
The default palettes are changed from 256 colors to 16, and the RGB
values are baked into the module.

Cubehelix now takes a sequence of RGB colors, like the other Palettable
subclasses. The `make` class method does the cubehelix algorithmic
calculations to generate a Cubehelix instance.

Also changed from OrderedDict to regular dictionaries with a palette_names
list for ordering. This fixes Python 2.6 compatibility.