cubehelix.get_map() does not use n arg
All the pre-built maps use 16 colors, and I'm not sure it makes sense to
build maps with different numbers of colors. Thus get_map() has no
explicit n keyword arg. But so that it won't fail I simply print out the
the user arguments there were supplied but unused. This will let the API
gracefully change in the future.